refactor(buildah): remove unused docker-with-fuse backend
